requirements.txt
python-telegram-bot==20.7
google-generativeai==0.7.0
python-dotenv==1.0.0
gspread==6.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0